chore(VO) : auth/vo 패키지로 이동
