--- src/main/kotlin/com/terning/server/kotlin/domain/auth/AuthType.kt
@@ -1,6 +1,0 @@
-package com.terning.server.kotlin.domain.auth
-
-enum class AuthType {
-    KAKAO,
-    APPLE,
-}
