+++ src/main/kotlin/com/terning/server/kotlin/domain/auth/vo/AuthId.kt
@@ -1,0 +1,8 @@
+package com.terning.server.kotlin.domain.auth.vo
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+data class AuthId(
+    val value: String,
+)
