+++ src/main/kotlin/com/terning/server/kotlin/domain/auth/vo/AuthType.kt
@@ -1,0 +1,6 @@
+package com.terning.server.kotlin.domain.auth.vo
+
+enum class AuthType {
+    KAKAO,
+    APPLE,
+}
