+++ src/main/kotlin/com/terning/server/kotlin/domain/auth/vo/RefreshToken.kt
@@ -1,0 +1,8 @@
+package com.terning.server.kotlin.domain.auth.vo
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+data class RefreshToken(
+    val value: String?,
+)
